image-upload: Fail hard if S3 credentials are missing
When uploading to multiple stores and one fails due to a network error,
this is tolerable -- so image-upload succeeds in that case.

But for the last months our CI machines were lacking the S3 token for
the local minio image cache on rhos-01-1, which is both systemic (and we
need to know about it!) and also expensive: Each of our ~ 40 bots will
re-download images from the primary Linode S3 store instead of the local
cache.