machine: add a lower-level rsync helper
Add a .rsync() command for running arbitrary rsync commands with the
test machine.  Use the RSYNC_CONNECT_COMMAND mechanism to invoke ssh,
allowing us to encode all knowledge of the user/address/port of the
remote machine into the wrapper: the caller can just say ":" to refer to
the machine.

Reimplement .upload(), .download(), .download_dir() in terms of the new
helper.  In addition to removing a whole lot of redundancy this also
removes all ssh-related knowledge from those commands.