machine: add a lower-level scp helper
Add an ._scp() helper for running arbitrary scp transfers with the test
machine.  Use "-o Hostname" to fill in the hostname, meaning that the
caller of this function doesn't need any knowledge of the ssh
destination and can just refer to it using something like "vm:".

Reimplement .upload(), .download(), .download_dir() in terms of the new
helper.  In addition to removing a whole lot of redundancy this also
removes all ssh-related knowledge from those commands.

This effectively removes our temporary use of rsync which helped us
avoid a bug in older sftp server implementations.  scp is dead as a
protocol, but there are no plans to eliminate the scp command (which
will live on as an alternate CLI for sftp), so we can safely use that
now.