stores: designate one of our S3 buckets "private"
Our attempts to mitigate scraper bots by restricting all access to
images is causing significant grief for new contributors: having to
privately message a few designated members of the Cockpit team in order
to get an S3 token to be able to start testing your first contribution
isn't particularly friendly or scalable.

One solution would be to generate a well-known "public" secret key for
downloading images.  The idea is that the scraper bots would not be
smart enough to use this key, but casual contributors would still be
able to come across it easily enough.  One issue with this approach is
that Linode's S3 ACLs are not sufficiently powerful to allow us to
specify that this key can only download the public (read: not RHEL)
images: we can only control access at the bucket level.

As a workaround, we can take advantage of the fact that we currently
have two more-or-less equal S3 buckets and arbitrarily designate the EU
one as "private" and the US one as "public".  We can upload the RHEL
images to only the "private" EU mirror and then give "public" access to
the US mirror for all the other images.

What happens with the new "public" secret key is currently up for
debate.  For the time being, let's keep it as a well-known secret passed
around Cockpit developers and shared freely with anyone who asks.  At
some point we may also consider to publish it in a README or even
integrate it into the downloader script directly.