images/scripts: add debian-trixie
Instead of doing the usual debian-stable → debian-stable dance, let's do
it more like Fedora and move to naming our Debian images by the release.

This way we can keep the old image in place while we're fixing tests in
the various projects, avoiding the need to land n+1 PRs in lockstep as
the definition of "debian-stable" changes everywhere at once.

Add the new image to "_manual" for the relevant projects: we can now
start fixing tests to work with "debian-trixie", updating the testmap as
we go.  Once we're done, we can drop the old "debian-stable" image.