images: patch selinux bugs in fedora-43 and -rawhide
Downstream bug https://bugzilla.redhat.com/show_bug.cgi?id=2374928 is
preventing us from using `test.thing` with our images.  sshd is properly
configured, but selinux is blocking it from accessing the ephemeral ssh
key that test.thing injects into the image.

Add a workaround for that.

This reveals another issue: ssh doesn't like the permissions of / which
are `rwxrwxr-x` for some reason.  Remove the group writeablity there.

These workarounds are only applied to fedora-43 and fedora-rawhide
because fedora-42 (and earlier) is missing the ephemeral key support.
Downstream plans a backport if/when the selinux issues (which we work
around here) are resolved.