jira-tasks: Add
Quick claude hackjob to create the "rebase to latest version" Jira
meta-issues that we need to upload into new RHEL releases. That's the
`rhel-rebase-issue` verb.

Also add a `find-field` verb to translate a UI custom field like
"Preliminary Testing" to an API name like `customfield_123abc`. That is
useful to automate the setting of these fields.