images: Move from Bitnami to official Grafana container image
Public Bitnami images will be discontinued at the end of the month [1].
In the meantime, the official Grafana image also learned how to inject
configuration, so we can move to that. The configuration looks a little
different now though, focusing on environment variables instead of an injected
ini file.

The first and second boot container invocation is now almost the same, except
for the enablement of the PCP plugin -- that can't yet happen on first boot, as
at that time the plugin isn't installed yet. But it's similar enough to reduce
the `podman` call to just the `run-grafana` script and special-case first boot.

We still have the problem of not being able to pull Dockerhub images
inside the RH VPN, so pick a reasonably looking quay.io mirror.

[1] https://hub.docker.com/r/bitnami/grafana/