images: work around ssh's "UNKNOWN" DNS lookups
Fedora and Debian already carry patches to help mitigate this, but Arch
and OpenSUSE are affected: when connecting from vhost, ssh incorrectly
tells PAM and libaudit that the remote host is called "UNKNOWN".  Both
of those components then attempt to perform lookups on this address and
fail.

We can work around that in our images by adding an `/etc/hosts` entry
for UNKNOWN to shortcut the lookup.  My first impulse was to use `::`
here but `/etc/hosts` is also used for reverse DNS lookups and this
might change the meaning of `::` for somebody, so let's use a completely
bogus address that would never appear anywhere else.