lib/aio: use subject.forge instead of ctx.forge
Our "Subject" object already contains reference to the "Forge" that it
is located on and uses this reference for computing things like its
clone url, but we ignore that field for some other operations and go
back up to the global JobContext.

That's going to make things more difficult when we have the possibility
of multiple forges.

Always use `Subject.forge` where possible.

8ecef463cd76 ("lib: port read_file over to rest api") removed support
for the separate `content-url` in the configuration but left a few
traces of it remaining, including a now-unused `.content_url` property
on `Subject`.  Change that to a `read_file()` method that calls the
`read_file()` method on the `forge` of the `Subject` and use that
instead of another `ctx.forge` reference.