job-runner: add support for multiple forges
Rename the `_forge: Forge` attribute on `JobContext` to `_forges` and
make it a dictionary.

Do a bit of retcon in job-runner.toml: unlike for logs, the names of the
`[forge.*]` sections no longer refer to specific drivers but are
identifiers.  Instead of selecting a `driver=` at the top-level, each of
these sections selects its own `driver=` and the toplevel now identifies
only a `default=` section by its identifier.

This should be more or less compatible with existing users: the only
valid section before was `[forge.github]` and this is the new default
value.  Meanwhile, the merging done between the `job-runner.toml` from
the user and our own default values will ensure that the `driver=` in
that section will be properly filled in.

Add a 'forge' field to the `SubjectSpecification` and a corresponding
`--forge` command-line argument.  If specified, this can be used to
override the default forge selection when resolving subjects.

Rename lib.aio.jobcontext.FORGES to FORGE_DRIVERS to match LOG_DRIVERS
and emphasize the point that this isn't about individual forges, but
rather about the API driver (of which there remains one: GitHub).

Add codeberg to the config.