tests-scan: merge some `if` blocks
The only way `project != api.repo` is if we `project = context_project`
in the previous block, so move it into that block.

Ditto: the only way `base != branch` is the `branch =` assigment in the
same block, so move it into the block as well and simplify its condition
by removing the `base != branch` check.