tests-scan: further simplification
We can plainly see that `branch` will never be `None` here (since it
falls back to the default branch in the assignment two lines up).  I
also suspect that `base is not None` is a bug, but we'll only ever hit
this bug if we do a cross-project test trigger on a non-default branch
on a raw sha in the bots repository, which has apparently never come up.

In any case: we definitely always want to checkout the branch specified
in the context, regardless of where the status is going...

With the condition now completely removed, we notice that we're
unconditionally reassigning checkout_ref right after it was assigned, so
remove the first assignment now.