tests-scan: add `@forge:` syntax
When specifying command subjects for test runs (ie: triggering
scenario@owner/repo from bots PRs) we now support specifying that the
repo is on a different forge by way of a `forge:` prefix.

It's now possible to trigger eg. `arch@codeberg:lis/test.thing` as a
context on a bots PR.