lib/aio: read token from ~/.config/cockpit-dev
If `post = false`, then according to the name of the forge, read the
token from the customary location.  For the default forge ('github')
this will do the right thing (~/.config/cockpit-dev/github-token), but
it also enables the possibility of using `codeberg-token`, for example.

This has the effect of enabling job-runner to work by default on most
developer machines, with posting of statuses and comments disabled by
default.