used-tasks-tags: rework, based on JobContext
This is already using the job-runner internals to do its work, so modify
it in the same way that we recently modified job-runner proper: we use
the JobContext to resolve the subject (supporting multiple forges) and
also adjust our .read_file() call to use the subject directly.

This removes any forge-specific knowledge from the script and makes it
work with the `codeberg:lis/test.thing` entry in the testmap.