image-customize: Support -node release tarballs in Debian source packages
When there is a project-node-version.tar.xz next to the
project-version.tar.xz upstream release tarball [1], include it into the
Debian source package as additional tarball. That will enable the Debian
package build to build the dist/ bundle from sources.

There's a gotcha here: We want the directory to be `node_modules/`, but
additional source tarball names can't have `_` in the name. So we call
and unpack it into `node`, and leave it to debian/rules to create a
symlink.

[1] our projects do that with `--upload $(NODE_CACHE):/var/tmp/`